+++ R/checkr-package.R
@@ -1,2 +1,3 @@
+#' @seealso \code{\link{check_vector}}, \code{\link{check_list}} and \code{\link{check_data}} 
 #' @keywords internal
 "_PACKAGE"