+++ R/join.R
@@ -1,9 +1,8 @@
 #' Check Join
 #' 
 #' Checks that the columns in data frame x form a many-to-one
-#' join with the corresponding columns in y.
-#' That is to say the join is a unique key in y and all the rows in
-#' x have a match in y.
+#' join with the corresponding columns in y, ie, 
+#' the join is a unique key in y and all the rows in x have a match in y.
 #'
 #' @param x The object to check.
 #' @param y The parent data frame.
@@ -13,6 +12,7 @@
 #' @param y_name A string of the name of the object y.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_data}}
 #' @export
 #' 
 #' @examples