+++ R/scalar.R
@@ -1,9 +1,12 @@
 #' Check Scalar
+#' 
+#' Checks whether an object is an atomic vector with one element.
 #'
 #' @param x The object to check.
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_vector}}
 #' @export
 #'
 check_scalar <- function(x,