+++ R/sorted.R
@@ -1,13 +1,12 @@
 #' Check Sorted
 #' 
-#' Checks whether object x is sorted.
-#' 
-#' Uses \code{!is.unsorted(x, na.rm = TRUE)} to test.
+#' Checks whether object x is sorted using \code{!is.unsorted(x, na.rm = TRUE)}.
 #' 
 #' @param x The object to check.
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_vector}} and \code{\link{check_list}}
 #' @export
 #'
 #' @examples
@@ -20,7 +19,8 @@ check_sorted <- function(x,
 
   check_flag_internal(error)
 
-  if (is.unsorted(x, na.rm = TRUE)) {
+  is_unsorted <- is.unsorted(x, na.rm = TRUE)
+  if (is.na(is_unsorted) || is_unsorted) {
     on_fail(x_name, " must be sorted", error = error)
   }
   invisible(x)