+++ R/string.R
@@ -7,6 +7,7 @@
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
+#' @seealso \code{\link{check_scalar}}
 #' @export
 #'
 #' @examples