+++ R/utils.R
@@ -76,18 +76,15 @@ tz <- function(x) {
   tz
 }
 
-check_class_internal <- function(x,
-                         values,
-                         x_name = substitute(x),
-                         error = TRUE) {
-  x_name <- deparse_x_name(x_name)
-
-  check_flag_internal(error)
-
-  class <- class(values)[1]
-
-  if (!inherits(x, class)) {
-    on_fail(x_name, " must be class ", class, error = error)
-  }
-  invisible(x)
+try_check <- function(expr) {
+  try(eval(expr, envir = parent.frame(3)), silent = TRUE)
 }
+
+try_message <- function(x) {
+  x <- as.character(x)
+  x <- sub("^Error.*[:] ", "", x, perl = TRUE)
+  x <- sub("\n$", "", x, perl = TRUE)
+  x
+}
+
+is.try_error <- function(x) inherits(x, "try-error")